+++ checkpoint/orbax/checkpoint/experimental/v1/_src/layout/registry.py
@@ -22,11 +22,13 @@ from orbax.checkpoint.experimental.v1._src.layout import safetensors_layout
 from orbax.checkpoint.experimental.v1._src.path import types as path_types
 
 InvalidLayoutError = checkpoint_layout.InvalidLayoutError
+CheckpointLayout = checkpoint_layout.CheckpointLayout
+CheckpointLayoutEnum = options_lib.CheckpointLayout
 
 
 async def get_checkpoint_layout(
-    path: path_types.PathLike, layout_enum: options_lib.CheckpointLayout
-) -> checkpoint_layout.CheckpointLayout:
+    path: path_types.PathLike, layout_enum: CheckpointLayoutEnum
+) -> CheckpointLayout:
   """Returns the checkpoint layout class for the given path.
 
   Args:
@@ -43,9 +45,9 @@ async def get_checkpoint_layout(
   path = epath.Path(path)
 
   match layout_enum:
-    case options_lib.CheckpointLayout.ORBAX:
+    case CheckpointLayoutEnum.ORBAX:
       layout_class = orbax_layout.OrbaxLayout
-    case options_lib.CheckpointLayout.SAFETENSORS:
+    case CheckpointLayoutEnum.SAFETENSORS:
       layout_class = safetensors_layout.SafetensorsLayout
     case _:
       raise ValueError(f"Unsupported checkpoint layout: {layout_enum}")
@@ -58,6 +60,82 @@ async def get_checkpoint_layout(
     raise InvalidLayoutError(
         f"Could not recognize the checkpoint at {path} as a valid"
         f" {layout_enum.value} checkpoint. If you are trying to load a"
-        " checkpoint that does not conform to the stadnard Orbax format, use"
+        " checkpoint that does not conform to the standard Orbax format, use"
         " `ocp.Context(layout=...)` to specify the expected checkpoint layout."
     ) from e
+
+
+async def get_checkpoint_layout_pytree(
+    path: path_types.PathLike,
+    layout_enum: CheckpointLayoutEnum,
+    checkpointable_name: str | None = None,
+) -> tuple[checkpoint_layout.CheckpointLayout, str | None]:
+  """Returns the checkpoint layout and checkpointable name for the given path."""
+  layout = await get_checkpoint_layout(path, layout_enum)
+  layout, checkpointable_name = (
+      await _try_resolve_pytree_checkpointable(
+          layout, checkpointable_name
+      )
+  )
+  await layout.validate_pytree(checkpointable_name)
+  return layout, checkpointable_name
+
+
+async def _try_resolve_pytree_checkpointable(
+    layout: CheckpointLayout,
+    checkpointable_name: str | None,
+) -> tuple[CheckpointLayout, str | None]:
+  """Tries to resolve the PyTree checkpointable name for a given layout.
+
+  Args:
+    layout: The CheckpointLayout object.
+    checkpointable_name: An optional name for the PyTree checkpointable.
+
+  Returns:
+    A tuple containing the (potentially updated) CheckpointLayout and the
+    resolved checkpointable name.
+
+  Raises:
+    ValueError: If it's a V0 checkpoint and a PyTree checkpointable name
+      cannot be resolved.
+  """
+  # Selected a specific name; use it.
+  if checkpointable_name is not None:
+    return layout, checkpointable_name
+  # Not a v0 checkpoint; use the default name.
+  if not _is_v0_checkpoint(layout):
+    return layout, checkpointable_name
+  # If it's a V0 checkpoint, we can try to resolve the checkpointable name from
+  # the path.
+  if not isinstance(layout, orbax_layout.OrbaxLayout):
+    raise AssertionError(f"Expected an OrbaxLayout, but got a {type(layout)}.")
+  # Option 1: It may be a direct path to the PyTree checkpointable.
+  try:
+    original_path = layout.path
+    new_layout = orbax_layout.OrbaxLayout(original_path.parent)
+    await new_layout.validate_pytree(original_path.name)
+    return new_layout, original_path.name
+  except checkpoint_layout.InvalidLayoutError:
+    pass
+  # Option 2: It may be a V0 checkpoint containing a PyTree checkpointable. It
+  # is possible for there to be multiple, but this would be unusual, and it is
+  # fine to just return the first one.
+  dir_names = [p.name for p in layout.path.iterdir() if p.is_dir()]
+  for name in dir_names:
+    try:
+      await layout.validate_pytree(name)
+    except checkpoint_layout.InvalidLayoutError:
+      continue
+    return layout, name
+  raise checkpoint_layout.InvalidLayoutError(
+      f"Detected an Orbax V0 checkpoint at {layout.path}, but failed to resolve"
+      " a checkpointable name for the PyTree checkpointable. Found"
+      f" subdirectory names: {dir_names}."
+  )
+
+
+def _is_v0_checkpoint(layout: CheckpointLayout) -> bool:
+  return not isinstance(layout, orbax_layout.OrbaxLayout) or (
+      isinstance(layout, orbax_layout.OrbaxLayout)
+      and not layout.has_indicator_file
+  )